requirements.txt
streamlit
langchain-google-genai
langchain
langchain-core
python-dotenv
pymongo
sentence-transformers
spacy
PyMuPDF